coordinat of Point now is double

+++ code/CCoA_ij_idea/src/Logic/FootprintSpaceTime/Point.java
@@ -1,10 +1,10 @@
 package Logic.FootprintSpaceTime;
 
 public interface Point {
-    public int  getX();
-    public int  getY();
-    public void setX(int  x);
-    public void setY(int  y);
+    public double  getX();
+    public double  getY();
+    public void setX(double  x);
+    public void setY(double  y);
     public Point clone();
 
     public boolean isLeftRelative(Point startLine, Point endLine);
